+++ frontend/src/components/navbar/LanguageSwitch.tsx
@@ -1,30 +1,34 @@
-import { useState } from "react"
-import { useTranslation } from "react-i18next"
-import { useNavigate } from "react-router-dom"
-import { LANGUAGES } from "@/locales/i18n"
-import { Dropdown } from "react-bootstrap"
-import "flag-icon-css/css/flag-icons.min.css"
-import "@/styles/navbar/LanguageSwitch.css"
+import { useState } from "react";
+import { useTranslation } from "react-i18next";
+import { useNavigate } from "react-router-dom";
+import { LANGUAGES } from "@/locales/i18n";
+import { Dropdown } from "react-bootstrap";
+import "flag-icon-css/css/flag-icons.min.css";
+import "@/styles/navbar/LanguageSwitch.css";
 
-const LanguageSwitch = () => {
-  const [isOpen, setIsOpen] = useState(false)
-  const { i18n } = useTranslation()
-  const navigate = useNavigate()
+interface LanguageSwitchProps {
+  onCollapse?: () => void; // Optional prop to handle navbar collapse
+}
+
+const LanguageSwitch: React.FC<LanguageSwitchProps> = ({ onCollapse }) => {
+  const [isOpen, setIsOpen] = useState(false);
+  const { i18n } = useTranslation();
+  const navigate = useNavigate();
 
   // Function to get the current language's flag icon
   const getCurrentFlag = () => {
-    const currentLang = LANGUAGES.find((lang) => lang.code === i18n.language)
-    return currentLang?.flag || "tr"
-  }
+    const currentLang = LANGUAGES.find((lang) => lang.code === i18n.language);
+    return currentLang?.flag || "tr";
+  };
 
   // Function to change the language and update the URL
-  const changeLang = (langCode: string) => {
-    i18n.changeLanguage(langCode) // Change the language in i18n
-    // Update URL to match new language
-    const currentPath = window.location.hash.split("/").slice(2).join("/")
-    navigate(`/${langCode}/${currentPath}`)
-    setIsOpen(false) // Close the dropdown
-  }
+  const changeLang = (langCode: string | undefined) => {
+    i18n.changeLanguage(langCode); // Change the language in i18n
+    const currentPath = window.location.hash.split("/").slice(2).join("/"); // Get the current path
+    navigate(`/${langCode}/${currentPath}`); // Navigate to the new language path
+    setIsOpen(false); // Close the dropdown
+    if (onCollapse) onCollapse(); // Collapse navbar if onCollapse prop is provided
+  };
 
   return (
     <Dropdown className="m-2" show={isOpen} onToggle={() => setIsOpen(!isOpen)}>
@@ -56,7 +60,7 @@ const LanguageSwitch = () => {
         ))}
       </Dropdown.Menu>
     </Dropdown>
-  )
-}
+  );
+};
 
-export default LanguageSwitch
+export default LanguageSwitch;