+++ frontend/src/components/navbar/Navbar.tsx
@@ -5,12 +5,13 @@ import CartIcon from "./CartIcon"
 import Logo from "./Logo"
 import Link from "./NavLink"
 import LanguageSwitch from "./LanguageSwitch"
-import "@/styles/navbar/Navbar.css"
+import "@/styles/navbar/navbar.css"
 
 const CustomNavbar = () => {
   const { t } = useTranslation()
   const [isExpanded, setIsExpanded] = useState(false)
 
+  // Navigation items
   const navItems = [
     { name: "navbar.home", href: "home" },
     { name: "navbar.about", href: "about" },
@@ -20,27 +21,29 @@ const CustomNavbar = () => {
     { name: "navbar.contact", href: "contact" },
   ]
 
+  // Function to collapse the navbar
+  const handleNavCollapse = () => setIsExpanded(false)
+
   return (
     <Navbar
       expand="lg"
-      className="mb-5 mx-5 px-5 custom-navbar"
+      className="mb-5 px-5"
       expanded={isExpanded}
-      sticky="top"
       onToggle={(expanded) => setIsExpanded(expanded)}
     >
       <Navbar.Brand>
         <Logo />
       </Navbar.Brand>
-      <Navbar.Toggle aria-controls="basic-navbar-nav" />
+      <Navbar.Toggle aria-controls="basic-navbar-nav" className="ms-auto" />
       <Navbar.Collapse id="basic-navbar-nav">
-        <Nav className="ms-auto mb-2" onClick={() => setIsExpanded(false)}>
+        <Nav className="ms-auto mb-2" onClick={handleNavCollapse}>
           {navItems.map((item) => (
             <Link key={item.name} name={t(item.name)} href={item.href} />
           ))}
         </Nav>
         <div className="d-flex justify-content-center align-items-center">
-          <CartIcon />
-          <LanguageSwitch />
+          <CartIcon onCollapse={handleNavCollapse} />
+          <LanguageSwitch onCollapse={handleNavCollapse} />
         </div>
       </Navbar.Collapse>
     </Navbar>