+++ frontend/src/pages/home/Home.tsx
@@ -2,7 +2,6 @@ import Carousel from "./Carousel"
 import Section from "./Section"
 import Partners from "../partners/Partners"
 import ProductCatalog from "../products/ProductCatalog"
-import Footer from "@/components/Footer"
 
 const Home = () => {
   return (
@@ -11,8 +10,6 @@ const Home = () => {
       <Section />
       <ProductCatalog showDetails={false} />
       <Partners />
-      <br />
-      <Footer />
     </>
   )
 }