+++ frontend/src/pages/products/CatalogPage.tsx
@@ -1,8 +1,7 @@
 import { useEffect } from "react"
-import Footer from "@/components/Footer"
 import ProductCatalog from "./ProductCatalog"
 
-const CatalogParent = () => {
+const CatalogPage = () => {
   // Scroll to the top of the page when the component mounts
   useEffect(() => {
     window.scrollTo(0, 0)
@@ -11,12 +10,9 @@ const CatalogParent = () => {
   // Render the Catalog component
   return (
     <div className="d-flex flex-column" style={{ minHeight: "100vh" }}>
-      <main className="mb-5" style={{ flexGrow: 1 }}>
-        <ProductCatalog showDetails={true} />
-      </main>
-      <Footer />
+      <ProductCatalog showDetails={true} />
     </div>
   )
 }
 
-export default CatalogParent
+export default CatalogPage